Removing the canvas module and moving everything into the renderer
