+++ src/turtle_window.rs
@@ -8,7 +8,7 @@ use std::sync::mpsc;
 use piston_window::math;
 
 use client;
-use canvas;
+use renderer;
 use animation::{Animation, MoveAnimation, RotateAnimation, AnimationStatus};
 use state::{TurtleState, DrawingState, Path};
 use query::{Query, Request, StateUpdate, DrawingCommand, Response};
@@ -50,7 +50,7 @@ impl TurtleWindow {
         // This needs to be called as close to the start of the program as possible
         // Since Turtle::new() is called at the beginning of many turtle programs, we do so here
         // to make sure this call occurs when it should.
-        canvas::setup();
+        renderer::setup();
 
         let (response_tx, response_rx) = mpsc::channel();
         let (renderer_process, handle) = renderer_client(response_tx);
